Replace IpcSynchronization with RunLoopSpinningWaiter
There was simple run loop spinning in `IpcSynchronization`, and there were some bugs probably with run loop spinning.

I replaced `IpcSynchronization` with code that uses `RunLoopSpinningWaiter`. `IpcSynchronization` was used in extension to `IpcClient` and `KeyboardEventInjector`.

I made `SynchronousIpcClient` and `SynchronousKeyboardEventInjector` with injected `RunLoopSpinningWaiter`. Because `SynchronousIpcClient` was used not only in tests, but also in `MixboxInAppServices`, I shared this code in `MixboxFoundation` (and that is not good, but it was just a place that is used by every other framework and I didn't want to make more frameworks).