Implement universal observing of UIWindow's UI events.
## UiEventObservingWindow

`UiEventObservingWindow` is now a part of framework (previously there was `TouchDrawingWindow` in tests).

It doesn't have any logic other than observing of events. All logic is implemented in observers, which can be dynamically added.

Moreover, a user of framework doesn't have to use `UiEventObservingWindow` to get logic of drawing touches or UIEvent history tracking. A user only have to register `UiEventObservable` in DI, for example, to use his own window.

There is a sugar added for registering a custom window:

`inAppServices.set(uiEventObservable: window)`

## TouchDrawer

`TouchDrawer` is now a part of frameworks. It's registered in default DI and is a `UiEventObserver`. It is disabled by default and can be enabled by calling `SetTouchDrawerEnabledIpcMethod`.

Example:

```
synchronousIpcClient
    .callOrFail(method: SetTouchDrawerEnabledIpcMethod(), arguments: true)
    .getVoidReturnValueOrFail()
```

It draws red circles for every touch, circles fade out in 250ms. It is very useful when watching what tests do.

## UiEventHistoryTracker

`UiEventHistoryTracker` is now a part of frameworks. It's registered in default DI and is a `UiEventObserver`. It is disabled by default and can be enabled by calling `SetUiEventHistoryTrackerEnabledIpcMethod`.

It records history of received UIEvent's. There is no practical need for it right now (for a user of Mixbox), but it can be used later in a helper that generates page object element's code for a view that user touches.

## Customizability of UiEventObserver's

Steps to add custom observer:

- Implement `UiEventObserver`
- Add IPC method for toggling the feature (use `BaseSetUiEventObserverEnabledIpcMethod`):

  ```
  public final class SetMyFeatureEnabledIpcMethod: BaseSetUiEventObserverEnabledIpcMethod {}
  ```

- Register IpcMethodHandler (use `UiEventObserverFeatureToggleIpcMethodHandler`):

  ```
  router.register(
      methodHandler: UiEventObserverFeatureToggleIpcMethodHandler(
          method: SetMyFeatureEnabledIpcMethod(),
          uiEventObserverFeatureToggleValueSetter: UiEventObserverFeatureToggler(
              uiEventObservableProvider: try dependencyResolver.resolve(),
              uiEventObserver: MyFeature()
          )
      )
  )
  ```

## Additional changes:
- `InAppServices` => `InAppServicesImpl`, to make interface more clear for user.
- Improve handling `Void` type in `IpcThrowingFunctionResult`: there is no need now to return `IpcVoid`, a static function `void` was added for this.
- `UnavoidableFailure.doOrFail` now supports injecting of file and line.
- `MultipleDependenciesRegistrationContinuation` now uses same scope for `reregister` as in `registerMultiple`
- `test___reregister` tests were added